@@ -1,14 +1,12 @@
 import 'dart:io';
 
-import 'package:flutter/foundation.dart';
-
 class PlaceLocation {
   final double latitude;
   final double longitude;
-  final String address;
+  final String? address;
 
   PlaceLocation(
-      {@required this.latitude, @required this.longitude, this.address});
+      {required this.latitude, required this.longitude, this.address});
 }
 
 class Place {
@@ -18,8 +16,8 @@ class Place {
   final File image;
 
   Place(
-      {@required this.id,
-      @required this.title,
-      @required this.image,
-      @required this.location});
+      {required this.id,
+      required this.title,
+      required this.image,
+      required this.location});
 }
